Update the token authentication & create exam

+++ onlineexam/OnlineExam/api/index.php
@@ -6,34 +6,40 @@ include 'functions.loader.php';
 
 $app = new Slim();
 
-$app->get('/test', function(){
-    $keys = ["No. of Items for Level 1", "No. of Items for Level 2",
-                "No. of Items for Level 3", "Passing Grade"];
+$app->get("/authenticate/:var", function($var) {
+    $param = explode(".", $var); //fsgggsgsgsg.4645475
+    $token = $param[0];
+    $uid = $param[1];
+    $verified = 0;
+    $error = 1;
+    $response = array();
+    if(count($param)===2) {
 
-    for($i=0;$i<count($keys);$i++) {
-        echo $keys[$i] . " ";
+        $verified = 1; $error = 0;
     }
+    $response = array(
+        "verified" => $verified,
+        "error" => $error
+    );
+    echo json_encode($response);
 });
 
-// $app->post("/sample/post/:var", function($var) {
-//     $param = explode(".", $var); //fsgggsgsgsg.4645475
-//     $token = $param[0];
-//     $uid = $param[1];
-//     $verified = 0;
-//     $error = 1;
-//     $response = array();
-//     if(count($param)===2) {
-
-//         $verified = 1; $error = 0;
-//     }
-//     $response = array(
-//         "verified" => $verified,
-//         "error" => $error
-//     );
-//     echo json_encode($response);
-// });
-
 //For registration of Exam Creator
+
+$app->get('/creatorDashboard/:var',function($var){
+    $param = explode(".", $var); 
+    $token = $param[0];
+    $uid = $param[1];
+    $verified = 0;
+    $error = 1;
+    $auth = checkToken($uid);
+    if(count($param)===2 && $auth) 
+    {
+        $getDashContents = dashboardData();
+        echo json_encode($getDashContents);
+    }
+    
+});
 $app->post('/creatorRegister',function(){
     $num = mt_rand(1000,9999);
     $str = substr(str_shuffle('1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ'),0,4);
@@ -107,11 +113,15 @@ $app->get('/examcode/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $examcode = date("Ymd").substr(str_shuffle('1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ'),0,6);
-        echo json_encode($examcode);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $examcode = date("Ymd").substr(str_shuffle('1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ'),0,6);
+            echo json_encode($examcode);
+        }
+        
     }
     
 });
@@ -123,20 +133,24 @@ $app->post('/addQuestion/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $keyIndex = ['question','choice1','choice2','choice3','choice4','answer'];
-        $examcode = $_POST['examcode'];
-        $level = $_POST['level'];
-        $examQuestion = $_POST['examQuestion'];
-        $choice1 = $_POST['choice1'];
-        $choice2 = $_POST['choice2'];
-        $choice3 = $_POST['choice3'];
-        $choice4 = $_POST['choice4'];
-        $answer = $_POST['answer'];
-        $insertQuestion = addQuestion($level,$examQuestion,$choice1,$choice2,$choice3,$choice4,$answer,$keyIndex,$examcode);
-        echo json_encode($insertQuestion);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $keyIndex = ['question','choice1','choice2','choice3','choice4','answer'];
+            $examcode = $_POST['examcode'];
+            $level = $_POST['level'];
+            $examQuestion = $_POST['examQuestion'];
+            $choice1 = $_POST['choice1'];
+            $choice2 = $_POST['choice2'];
+            $choice3 = $_POST['choice3'];
+            $choice4 = $_POST['choice4'];
+            $answer = $_POST['answer'];
+            $insertQuestion = addQuestion($level,$examQuestion,$choice1,$choice2,$choice3,$choice4,$answer,$keyIndex,$examcode);
+            echo json_encode($insertQuestion);
+        }
+        
     }
     
 });
@@ -148,15 +162,19 @@ $app->post('/createExam/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $examcode = $_POST['examcode'];
-        $category = $_POST['category'];
-        $subject = $_POST['subject'];
-        $level = $_POST['level'];
-        $createExam = addExam($category,$subject,$level,$examcode);
-        echo json_encode($createExam);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $examcode = $_POST['examcode'];
+            $category = $_POST['category'];
+            $subject = $_POST['subject'];
+            $level = $_POST['level'];
+            $createExam = addExam($category,$subject,$level,$examcode);
+            echo json_encode($createExam);
+        }
+        
     }
     
 });
@@ -168,11 +186,15 @@ $app->get('/viewCreatedExams/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $createdExams = viewCreatedExams();
-        echo json_encode($createdExams);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $createdExams = viewCreatedExams();
+            echo json_encode($createdExams);
+        }
+        
     }   
 
 });
@@ -183,14 +205,18 @@ $app->post('/deleteExam/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $delID = $_POST['delID'];
-        $delete = deleteExam($delID);
-        echo json_encode($delete);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $delID = $_POST['delID'];
+            $delete = deleteExam($delID);
+            echo json_encode($delete);
+        }
     }  
 });
+
 $app->post('/deleteDuplicatedExam/:var',function($var){
     $param = explode(".", $var); 
     $token = $param[0];
@@ -198,12 +224,16 @@ $app->post('/deleteDuplicatedExam/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $examcode = $_POST['examcode'];
-        $delete = deleteDuplicatedExam($examcode);
-        echo json_encode($delete);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $examcode = $_POST['examcode'];
+            $delete = deleteDuplicatedExam($examcode);
+            echo json_encode($delete);
+        }
+        
     }
     
 });
@@ -215,11 +245,14 @@ $app->get('/getCategory/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2)     
     {
-        $verified = 1; $error = 0;
-        $fetchCategory = getCategory();
-        echo json_encode ($fetchCategory);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $fetchCategory = getCategory();
+            echo json_encode ($fetchCategory);
+        }
     }
 });
 
@@ -231,13 +264,16 @@ $app->post('/addCategory/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $categoryCode = mt_rand(1000,9999);
-        $categoryName = $_POST['categoryName'];
-        $insertCategory = addCategory( $categoryCode,$categoryName);
-        echo json_encode($insertCategory);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $categoryCode = mt_rand(1000,9999);
+            $categoryName = $_POST['categoryName'];
+            $insertCategory = addCategory( $categoryCode,$categoryName);
+            echo json_encode($insertCategory);
+        }
     }
 });
 //Fetches category to edit
@@ -248,12 +284,16 @@ $app->post('/fetchCategoryEdit/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $editID = $_POST['editID'];
-        $fetch = getCategoryData($editID);
-        echo json_encode($fetch);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $editID = $_POST['editID'];
+            $fetch = getCategoryData($editID);
+            echo json_encode($fetch);
+        }
+       
     }
 });
 //Updating category
@@ -263,13 +303,17 @@ $app->post('/editCategory/:var',function($var){
     $uid = $param[1];
     $verified = 0;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth) 
+    if(count($param)===2) 
     {
-        $verified = 1; $error = 0;
-        $editID = $_POST['editID'];
-        $newCategory = $_POST['newCategory'];
-        $edit = editCategory($editID,$newCategory);
-        echo json_encode($edit);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $editID = $_POST['editID'];
+            $newCategory = $_POST['newCategory'];
+            $edit = editCategory($editID,$newCategory);
+            echo json_encode($edit);
+        }
+        
     }
 });
 //Deleting category
@@ -280,12 +324,16 @@ $app->post('/deleteCategory/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth)
+    if(count($param)===2)
     {
-        $verified = 1; $error = 0;
-        $delID = $_POST['delID'];
-        $delete = deleteCategory($delID);
-        echo json_encode($delete);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $delID = $_POST['delID'];
+            $delete = deleteCategory($delID);
+            echo json_encode($delete);
+        }
+        
     }
 });
 //Fetches Subject
@@ -296,11 +344,15 @@ $app->get('/getSubjects/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth)
+    if(count($param)===2)
     {
-        $verified = 1; $error = 0;
-        $fetchSubject = getSubject();
-        echo json_encode ($fetchSubject);
+        if($auth!=0)
+        {
+            $verified = 1; $error = 0;
+            $fetchSubject = getSubject();
+            echo json_encode ($fetchSubject);
+        }
+       
     }
 });
 //Creates new subject
@@ -311,12 +363,16 @@ $app->post('/addSubject/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth)
+    if(count($param)===2)
     {
-        $subjectCategory = $_POST['subjectCategory'];
-        $subjectName = $_POST['subjectName'];
-        $insertSubject = addSubject($subjectCategory,$subjectName);
-        echo json_encode($insertSubject);
+        if($auth!=0)
+        {
+            $subjectCategory = $_POST['subjectCategory'];
+            $subjectName = $_POST['subjectName'];
+            $insertSubject = addSubject($subjectCategory,$subjectName);
+            echo json_encode($insertSubject);
+        }
+        
     }
 });
 //Fetches subject to edit
@@ -326,11 +382,15 @@ $app->post('/getSubjectEdit/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth)
+    if(count($param)===2)
     {
-        $editID = $_POST['editID'];
-        $fetch = getSubjectData($editID);
-        echo json_encode($fetch);
+        if($auth!=0)
+        {
+            $editID = $_POST['editID'];
+            $fetch = getSubjectData($editID);
+            echo json_encode($fetch);
+        }
+        
     }
     
 });
@@ -342,13 +402,17 @@ $app->post('/editSubject/:var',function($var){
     $verified = 0;
     $error = 1;
     $auth = checkToken($uid);
-    if(count($param)===2 && $auth)
+    if(count($param)===2)
     {
-        $editID = $_POST['editID'];
-        $uid = $_POST['uid'];
-        $newSubject = $_POST['newSubject'];
-        $edit = editSubject($editID,$uid,$newSubject);
-        echo json_encode($edit);
+        if($auth!=0)
+        {
+            $editID = $_POST['editID'];
+            $uid = $_POST['uid'];
+            $newSubject = $_POST['newSubject'];
+            $edit = editSubject($editID,$uid,$newSubject);
+            echo json_encode($edit);
+        }
+        
     }
 });
 //Deleting subject
@@ -361,9 +425,13 @@ $app->post('/deleteSubject/:var',function($var){
     $auth = checkToken($uid);
     if(count($param)===2 && $auth)
     {
-        $delID = $_POST['delID'];
-        $delete = deleteSubject($delID);
-        echo json_encode($delete);
+        if($auth!=0)
+        {
+            $delID = $_POST['delID'];
+            $delete = deleteSubject($delID);
+            echo json_encode($delete);
+        }
+        
     }
 });
 
@@ -387,8 +455,12 @@ $app->get('/getExaminees/:var',function($var){
     $auth = checkToken($uid);
     if(count($param)===2 && $auth)
     {
-        $fetchExaminees = getExaminees();
-        echo json_encode($fetchExaminees);
+        if($auth!=0)
+        {
+            $fetchExaminees = getExaminees();
+            echo json_encode($fetchExaminees);
+        }
+        
     }
     
 });
